requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
psutil>=5.9.0

# Audio processing
librosa>=0.9.0
soundfile>=0.11.0
audioread>=3.0.0
demucs>=4.0.0
pedalboard>=0.7.0

# GUI
gradio>=4.0.0
matplotlib>=3.5.0

# Deep learning (choose one or both based on your needs)
torch>=2.0.0
torchaudio>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
colorlog>=6.7.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: for advanced features
matplotlib>=3.5.0
scikit-learn>=1.0.0
